+++ src/types/api.ts
@@ -1,3 +1,0 @@
-export interface GetItemsOptions {
-	showProgress?: boolean
-}