+++ CHANGELOG.md
@@ -5,9 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [5.0.0](../../tags/v5.0.0) - 2023-03-20
+## [6.0.0](../../tags/v6.0.0) - 2023-03-20
 ### Added
 - Non-persistence mode for getAuth: ability to not save sensitive credentials into the file
+- `hideProgress` option for `login` that is false by default
+
+### Changed
+- `showProgress` changed to `hideProgress` and it's false by default
+- `persist` changed to `temporary` and it's false by default
 
 ## Removed
 - Removed getter methods from api helpers to be able to re-use auth between api usages. Now better call `getAPI` and then run needed native methods on their own.